% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisper_install.R
\name{get_whisper_version}
\alias{get_whisper_version}
\title{Check the version of Whisper}
\usage{
get_whisper_version(envname = NULL)
}
\arguments{
\item{envname}{The name of, or path to, a Python virtual environment. Default
is to use "r-whisper" unless set in environment variable WHISPER_PYTHON
(through \code{Sys.setenv()}).}
}
\value{
string with version of Whisper
}
\description{
Check the version of Whisper
}
