% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisper_install.R
\name{install_whisper}
\alias{install_whisper}
\title{Install OpenAI's Whisper plus dependencies}
\usage{
install_whisper(
  envname = NULL,
  whisper_version = NULL,
  gpu = FALSE,
  ask = TRUE
)
}
\arguments{
\item{envname}{The name of, or path to, a Python virtual environment. Default
is to use "r-whisper" unless set in environment variable WHISPER_PYTHON
(through \code{Sys.setenv()}).}

\item{whisper_version}{give a specific version if you want.}

\item{gpu}{install packages needed for GPU usage.}

\item{ask}{set \code{FALSE} for unattended install.}
}
\value{
\code{TRUE} if everything worked.
}
\description{
Install OpenAI's Whisper plus dependencies
}
