% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whisper.R
\name{transcribe}
\alias{transcribe}
\title{Transcribe an audio file}
\usage{
transcribe(
  audio,
  model = "base",
  device = getOption("CUDA_AVAILABLE", default = "cpu"),
  background = TRUE,
  envname = NULL,
  verbose = TRUE
)
}
\arguments{
\item{audio}{path to a file that contains audio (e.g., .mp3, .mp4).}

\item{model}{name of the model. One of "tiny", "base", "small", "medium", or
"large". Or an English-only models, which have the same names but with the
suffix ".en" (e.g. "tiny.en").}

\item{device}{"cpu" or "cuda" to run transcription on the GPU (which is
faster, if you have one).}

\item{background}{run in a background process?}

\item{envname}{name of the virtual environment used for whisperrr.}

\item{verbose}{whether status messages should be printed to screen.}
}
\value{
as data.frame with transcriptions
}
\description{
Transcribe an audio file
}
\examples{
\dontrun{
transcription <- transcribe("audio.mp3")
}
}
